Much less jittery camera.

+++ src/game_engine_stuff/level_stuff/active_level_manager_class.arm.cpp
@@ -153,13 +153,18 @@ scr_entry active_level_manager::horiz_sublevel_tile_at( u32 tile_x,
 void active_level_manager::correct_bg0_scroll_mirror 
 	( const vec2_u32& sublevel_size_in_blocks_2d )
 {
-	//s16 sublevel_x_coord_tile = gfx_manager::bgofs_mirror[0].curr.x >> 3;
-	//s16 sublevel_y_coord_tile = gfx_manager::bgofs_mirror[0].curr.y >> 3;
+	//s32 sublevel_x_coord_tile = gfx_manager::bgofs_mirror[0].curr.x >> 3;
+	//s32 sublevel_y_coord_tile = gfx_manager::bgofs_mirror[0].curr.y >> 3;
 	
-	s16 sublevel_x_coord_tile = gfx_manager::bgofs_mirror[0].curr.x
+	s32 sublevel_x_coord_tile = gfx_manager::bgofs_mirror[0].curr.x
 		.trunc_to_int() >> 3;
-	s16 sublevel_y_coord_tile = gfx_manager::bgofs_mirror[0].curr.y
+	s32 sublevel_y_coord_tile = gfx_manager::bgofs_mirror[0].curr.y
 		.trunc_to_int() >> 3;
+	
+	//s32 sublevel_x_coord_tile = gfx_manager::bgofs_mirror[0].curr.x
+	//	.true_round_via_trunc() >> 3;
+	//s32 sublevel_y_coord_tile = gfx_manager::bgofs_mirror[0].curr.y
+	//	.true_round_via_trunc() >> 3;
 	
 	
 	if ( sublevel_x_coord_tile < 0 )
@@ -169,7 +174,7 @@ void active_level_manager::correct_bg0_scroll_mirror
 	}
 	
 	else if ( sublevel_x_coord_tile 
-		> (s16)(sublevel_width_in_tiles(sublevel_size_in_blocks_2d) 
+		> (s32)(sublevel_width_in_tiles(sublevel_size_in_blocks_2d) 
 		- screen_width_in_tiles - 1 ) )
 	{
 		//gfx_manager::bgofs_mirror[0].curr.x = ( sublevel_width_in_tiles
@@ -190,7 +195,7 @@ void active_level_manager::correct_bg0_scroll_mirror
 	}
 	
 	else if ( sublevel_y_coord_tile 
-		> (s16)(sublevel_height_in_tiles(sublevel_size_in_blocks_2d)
+		> (s32)(sublevel_height_in_tiles(sublevel_size_in_blocks_2d)
 		- screen_height_in_tiles -1 ) )
 	{
 		//gfx_manager::bgofs_mirror[0].curr.y = ( sublevel_height_in_tiles
@@ -214,11 +219,11 @@ void active_level_manager::update_sublevel_in_screenblock_mirror_2d()
 	
 	
 	// VRAM tiles are 8x8 pixels in size, so divide the 
-	//s16 sublevel_x_coord_tile = gfx_manager::bgofs_mirror[0].curr.x >> 3;
-	//s16 sublevel_y_coord_tile = gfx_manager::bgofs_mirror[0].curr.y >> 3;
-	s16 sublevel_x_coord_tile = gfx_manager::bgofs_mirror[0].curr.x
+	//s32 sublevel_x_coord_tile = gfx_manager::bgofs_mirror[0].curr.x >> 3;
+	//s32 sublevel_y_coord_tile = gfx_manager::bgofs_mirror[0].curr.y >> 3;
+	s32 sublevel_x_coord_tile = gfx_manager::bgofs_mirror[0].curr.x
 		.trunc_to_int() >> 3;
-	s16 sublevel_y_coord_tile = gfx_manager::bgofs_mirror[0].curr.y
+	s32 sublevel_y_coord_tile = gfx_manager::bgofs_mirror[0].curr.y
 		.trunc_to_int() >> 3;
